Assertion hit when saving to PDF from printing dialog https://bugs.webkit.org/show_bug.cgi?id=201351
Reviewed by Tim Horton.

[WKPrintingView knowsPageRange:] may get called from a background thread and ends up calling
WeakPtr::get() which asserts because it is generally unsafe to get the raw pointer from a
WeakPtr to an object owned by another thread. However, the methods here merely want to null
check it so add a thread-safe method (pageIsClosed()) to WebFrameProxy for it.

* UIProcess/WebFrameProxy.h:
(WebKit::WebFrameProxy::pageIsClosed const):
* UIProcess/mac/WKPrintingView.mm:
(-[WKPrintingView knowsPageRange:]):


Identifier: 215015@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@249346 268f45cc-cd09-0410-ab3c-d52691b4dbfc