REGRESSION (r249364): Layout Test webgpu/texture-triangle-strip.html is Failing https://bugs.webkit.org/show_bug.cgi?id=201425
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2019-09-04
Reviewed by Myles C. Maxfield.

The orientation of webgpu/resources/blue-checkered.png is OriginRightTop.
Because r249364 made the image orientation respected everywhere including
the canvas drawImage(), the expected result does not match the output of
the test anymore.

The fix is to change the the expected result to match the image drawing
when the image orientation is respected.

* webgpu/js/webgpu-functions.js:
(drawGreenAndBlueCheckerboardInSoftware):

Identifier: 215089@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@249460 268f45cc-cd09-0410-ab3c-d52691b4dbfc