[WebAuthn] Make WebAuthn default off and let clients turn it on at will https://bugs.webkit.org/show_bug.cgi?id=201439 <rdar://problem/54998154>
Reviewed by Youenn Fablet.

This patch makes WebAuthn default off such that clients that have the right entitlements
could turn it on and we don't risk at turning on a Web API that does nothing by default.

This patch doesn't add any SPI to turn the feature on as it is currently doable via
- [WKPreferencesPrivate _setEnabled:forExperimentalFeature:].

* Shared/WebPreferences.yaml:
* Shared/WebPreferencesDefaultValues.cpp:
(WebKit::defaultWebAuthenticationEnabled): Deleted.
* Shared/WebPreferencesDefaultValues.h:


Identifier: 215199@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@249600 268f45cc-cd09-0410-ab3c-d52691b4dbfc