Remove alternate declaration of CCECCryptorImportKey https://bugs.webkit.org/show_bug.cgi?id=201903 <rdar://problem/55466589>
Reviewed by Chris Dumez.

In macOS 10.15 and iOS 13.0 SDKs, the keyPackage parameter of
CCECCryptorImportKey() was changed from "void*" to "const void*". This
change was accounted for in Bug 195754 by making the same change to
the parallel declarations in CommonCryptoSPI.h. But since this header
also needed to be used in macOS 10.14 contexts, the old declaration
was also kept.

But since this function is a "C" function, the types of parameters are
not included in the function signature, and it doesn't really matter
what parameter types are, so long as the right values are sent to the
called function. This means that we don't need two versions of the
function, one with a "const" parameter and one without. We only really
need one. In this case, it makes sense to keep the more-restrictive
one -- the with a "const void*" parameter. Therefore, remove the one
taking just "void*".

* pal/spi/cocoa/CommonCryptoSPI.h:


Identifier: 215574@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250049 268f45cc-cd09-0410-ab3c-d52691b4dbfc